+++ crates/teloxide-core/src/adaptors/trace.rs
@@ -239,6 +239,7 @@ where
         create_invoice_link,
         answer_shipping_query,
         answer_pre_checkout_query,
+        refund_star_payment,
         set_passport_data_errors,
         send_game,
         set_game_score,